memcg: add mem_cgroup_mark_inode_dirty()
Create the mem_cgroup_mark_inode_dirty() routine, which is called when
an inode is marked dirty.  In kernels without memcg, this is an inline
no-op.

Add i_memcg field to struct address_space.  When an inode is marked
dirty with mem_cgroup_mark_inode_dirty(), the css_id of current memcg is
recorded in i_memcg.

The type of i_memcg is an 'unsigned short' because it stores the css_id
of the memcg.  Using a struct mem_cgroup pointer would be larger and
also create a reference on the memcg which would hang memcg rmdir
deletion.  Usage of a css_id is not a reference so cgroup deletion is
not affected.  The memcg can be deleted without cleaning up the i_memcg
field.  When a memcg is deleted its pages are recharged to the cgroup
parent, and the related inode(s) are marked as shared thus
disassociating the inodes from the deleted cgroup.

A special I_MEMCG_SHARED value (zero, which is an illegal css_id value)
is used in i_memcg for address spaces that cannot be attributed to
exactly one cgroup.  Address spaces shared by cgroups or ones dirtied
when the cgroup cannot be determined are marked with this shared value.
As inode pages are marked dirty, if the dirtied page's cgroup differs
from the address space's previously recorded i_memcg cgroup, then the
address space is marked shared.

Per-memcg writeback (introduced in a later change) uses this field to
isolate inodes associated with a particular memcg.

A mem_cgroup_mark_inode_dirty() tracepoint is also included to allow for
easier understanding of memcg writeback operation.

Signed-off-by: Greg Thelen <gthelen@google.com>