memcg: check memcg dirty limits in page writeback
If the current process is in a non-root memcg, then
balance_dirty_pages_ratelimited_nr() will consider the memcg dirty
limits as well as the system-wide limits.  This allows different cgroups
to have distinct dirty limits which trigger direct and background
writeback at different levels.

If called with a mem_cgroup, then throttle_vm_writeout() queries the
given cgroup for its dirty memory usage limits and only throttles if the
calling task may contribute writeback pages to the under-reclaim cgroup.
The rationale for this needs some explaining.
Consider memcg hierarchy:
        A
      A1 A2

Assume A1.dirty_limit=0.  If A2 induces global page reclaim (or direct
reclaim on A), then A2 will attempt to reclaim pages from A1 with
do_shrink_zone().  Once do_shrink_zone() finishes reclaiming it uses
throttle_vm_writeout() which would see that the number of
A1.nr_writeback pages exceeds the A1.dirty_limit.  If other tasks
continue to add dirty pages to A1, then writeback will be started
immediately due to the low A1.dirty_limit.  This continued insertion of
A1 writeback pages can cause livelock as A2 reclaim takes a very long
time to complete waiting for A1's nr_writeback to fall to zero.

Thus this patch only throttles a task if it may contribute writeback
pages to the under-reclaim cgroup.

Child cgroups are intentionally not checked.  This protects against a
different configuration where A1 and A2 both have large dirty limits and
A.dirty_limit=0.  In this case an aggressive writer in A1 would never
allow A.nr_writeback to hit zero for the same reasons described above.
A reclaiming task in A2 is thus protected by this change from a heavy
writer in A1.

The only case where throttling by throttle_vm_writeout() will cause a
non-dirtying task to sleep is if it shares a memcg with a dirtying task.
In this case whichever condition (either global or target reclaim)
causes the non-dirtying task to wait in throttle_vm_writeout() will also
cause the dirtying task(s) to also reclaim and similarly wait.

This change does unfortunately introduce a lockdep violation related to
usage of get_online_cpus() while in the reclaim path.  This is also
present in 6xx and tracked as b/8793093.  I have a possible fix for this
lockdep issue and would love to avoid introducing another lockdep issue.
But it's not something required for rebase and the fix is non-trivial,
so I'm not rushing the fix.  The call stack which triggers lockdep is:
  throttle_vm_writeout() ->
    mem_cgroup_hierarchical_dirty_info() ->
      mem_cgroup_read_stat() ->
        get_online_cpus()
The problem is that cpu init code grabs the cpu hotplug mutex and then
calls alloc_pages(GFP_KERNEL) which can reclaim and call
throttle_vm_writeout().  Such reclaim would deadlock on a nested call to
get_online_cpus().  I suspect the possibility of seeing a deadlock is
very low because would require memory reclaim during cpu initialization.
6xx kernels share this issue and have never seen this deadlock.

Signed-off-by: Greg Thelen <gthelen@gthelen.com>