memcg: check dirty ratios when limits are adjusted
When either the hard limit, soft limit, dirty ratios, or dirty limits
are adjusted, recheck the background dirty limits.  If the adjusted
memcg is now over the background limit then wakeup the flusher threads
to bring the memcg back into compliance.  The flusher threads write data
asynchronously, so the background dirty ratios may be exceed immediately
after change the limits.  The ratios should be reconciled soon as the
background flusher write data.

Added a /proc/writeback/stats counter to display the number of times
this new logic is used to wakeup flushers.

Signed-off-by: Greg Thelen <gthelen@google.com>